Added accounts app, added 1 api request
