+++ accounts/templates/signup.html
