+++ link_shortener/settings.py
@@ -45,6 +45,7 @@ INSTALLED_APPS = [
     "rest_framework.authtoken",
     "drf_spectacular",
     "shortener",
+    "accounts",
 ]
 
 MIDDLEWARE = [
